requirements.txt
paho-mqtt
opencv-python
mediapipe
sys
argparse
datetime